Next commit to Git. Modified export_goods package, changes code in views, signals, tasks, urls, models, settings, screens folder.

+++ orders/orders/tasks.py
@@ -1,8 +1,12 @@
-import yaml
-from orders.settings import EMAIL_HOST_USER
+from django.http import JsonResponse
 from django.core.mail.message import EmailMultiAlternatives
+from django.core.validators import URLValidator
+from django.core.exceptions import ValidationError
 from .celery_conf import app
+from requests import get
+from yaml import load as load_yaml, Loader
 from web_service.models import Shop, Category, Product, Parameter, ProductParameter, ProductInfo
+from orders.settings import DEFAULT_EMAIL_ADMIN, EMAIL_HOST_USER
 
 from celery.utils.log import get_task_logger
 
@@ -25,7 +29,7 @@ class FaultTolerantTask(celery.Task):
 
 @app.task(base=FaultTolerantTask, name='send_email_task')
 def send_emails(title, message, email, *args, **kwargs):
-    emails = []
+    emails = [DEFAULT_EMAIL_ADMIN]
     emails.append(email)
     try:
         msg = EmailMultiAlternatives(subject=title, body=message, from_email=EMAIL_HOST_USER, to=emails)
@@ -35,38 +39,52 @@ def send_emails(title, message, email, *args, **kwargs):
         raise ex
 
 
-def open_file(shop):
-    with open(shop.get_file(), 'r') as file:
-        data = yaml.safe_load(file)
-    return data
-
-
 @app.task(base=FaultTolerantTask, name='import_shop_task')
-def import_shop_data(data, user_id):
-    file = open_file(data)
+def import_shop_data(request):
+    if not request.user.is_authenticated:
+        return JsonResponse({'Status': False, 'Error': 'Log in required'}, status=403)
 
-    shop, _ = Shop.objects.get_or_create(user_id=user_id,
-                                         defaults={'name': file['shop']})
+    if request.user.type != 'shop':
+        return JsonResponse({'Status': False, 'Error': 'Только для магазинов'}, status=403)
 
-    for category in file['categories']:
-        category_object, _ = Category.objects.get_or_create(id=category['id'], name=category['name'])
-        category_object.shops.add(shop.id)
-        category_object.save()
+    url = request.data.get('url')
+    if url:
+        validate_url = URLValidator()
+        try:
+            validate_url(url)
+        except ValidationError as ex:
+            return JsonResponse({'Status': False, 'Error': str(ex)})
+        else:
+            stream = get(url).content
 
-    ProductInfo.objects.filter(shop_id=shop.id).delete()
+            data = load_yaml(stream, Loader=Loader)
 
-    for item in file['goods']:
-        product, _ = Product.objects.get_or_create(name=item['name'], category_id=item['category'])
+            shop, _ = Shop.objects.get_or_create(
+                name=data['shop'], user_id=request.user.id)
+            for category in data['categories']:
+                category_object, _ = Category.objects.get_or_create(
+                    id=category['id'], name=category['name'])
+                category_object.shops.add(shop.id)
+                category_object.save()
+            ProductInfo.objects.filter(shop_id=shop.id).delete()
+            for item in data['goods']:
+                product, _ = Product.objects.get_or_create(
+                    name=item['name'], category_id=item['category'])
 
-        product_info = ProductInfo.objects.create(product_id=product.id,
-                                                  external_id=item['id'],
-                                                  model=item['model'],
-                                                  price=item['price'],
-                                                  price_rrc=item['price_rrc'],
-                                                  quantity=item['quantity'],
-                                                  shop_id=shop.id)
-        for name, value in item['parameters'].items():
-            parameter_object, _ = Parameter.objects.get_or_create(name=name)
-            ProductParameter.objects.create(product_info_id=product_info.id,
-                                            parameter_id=parameter_object.id,
-                                            value=value)
+                product_info = ProductInfo.objects.create(product_id=product.id,
+                                                          external_id=item['id'],
+                                                          model=item['model'],
+                                                          price=item['price'],
+                                                          price_rrc=item['price_rrc'],
+                                                          quantity=item['quantity'],
+                                                          shop_id=shop.id)
+                for name, value in item['parameters'].items():
+                    parameter_object, _ = Parameter.objects.get_or_create(
+                        name=name)
+                    ProductParameter.objects.create(product_info_id=product_info.id,
+                                                    parameter_id=parameter_object.id,
+                                                    value=value)
+
+            return JsonResponse({'Status': True})
+
+    return JsonResponse({'Status': False, 'Errors': 'Не указаны все необходимые аргументы'})
